Refactoring and small code improvements
