+++ services-impl/services-area009-v001-mpd/src/main/java/esa/mo/services/mpd/consumer/ProductOrderDeliveryConsumerServiceImpl.java
@@ -49,7 +49,7 @@ public class ProductOrderDeliveryConsumerServiceImpl extends ConsumerServiceImpl
             throws MALException, MalformedURLException, MALInteractionException {
         this.connectionDetails = connectionDetails;
 
-        // Close old connection
+        // Close previous connection
         if (tmConsumer != null) {
             try {
                 tmConsumer.close();
